event jwt 적용 및 cors 적용

+++ routes/events.py
@@ -1,8 +1,9 @@
 from typing import List
 
 from beanie import PydanticObjectId
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
+from planner.auth.authenticate import authenticate
 from planner.database.connection import Database
 from planner.models.events import Event
 
@@ -34,7 +35,8 @@ async def get_event(id: PydanticObjectId) -> Event:
 
 
 @event_router.post("/")
-async def create_event(event: Event) -> dict:
+async def create_event(event: Event, user: str = Depends(authenticate)) -> dict:
+    event.creator = user
     await event_database.save(event)
     return {
         'message': 'event created successfully',
@@ -42,7 +44,13 @@ async def create_event(event: Event) -> dict:
 
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, event: Event) -> Event:
+async def update_event(id: PydanticObjectId, event: Event, user: str = Depends(authenticate)) -> Event:
+    event_exist: Event = await event_database.get(id)
+    if event_exist.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="You do not have permission to perform this action",
+        )
     updated_event = await event_database.update(id, event)
     if not updated_event:
         raise HTTPException(
@@ -53,7 +61,13 @@ async def update_event(id: PydanticObjectId, event: Event) -> Event:
 
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event_exist: Event = await event_database.get(id)
+    if event_exist.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="You do not have permission to perform this action",
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
@@ -62,12 +76,4 @@ async def delete_event(id: PydanticObjectId) -> dict:
         )
     return {
         'message': 'Event deleted',
-    }
-
-
-@event_router.delete("/")
-async def delete_all_events() -> dict:
-    events.clear()
-    return {
-        'message': 'All event deleted',
     }
